requirements.txt
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
zipp==3.15.0
absl-py==0.12.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
Appium-Python-Client==2.2.0
argh==0.26.2
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Automat==22.10.0
aws==0.2.5
awscli==1.20.29
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
boto==2.49.0
boto3==1.26.14
botocore==1.29.14
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
constantly==15.1.0
cryptography==37.0.1
distlib==0.3.2
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
docker==5.0.3
docutils==0.15.2
fabric==2.6.0
filelock==3.0.12
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flatbuffers==1.12
frozenlist==1.3.3
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.18.2
greenlet==2.0.2
grpcio==1.32.0
h11==0.13.0
h5py==2.10.0
humanize==0.5.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.3.0
importlib-resources==5.12.0
incremental==22.10.0
invoke==1.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
Keras-Preprocessing==1.1.2
linecache2==1.0.0
lxml==4.6.3
Mako==1.2.4
Markdown==2.6.5
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.19.4
oauthlib==3.2.0
openai==0.27.4
opt-einsum==3.3.0
outcome==1.1.0
paramiko==2.7.2
pathlib2==2.3.6
pathtools==0.1.2
platformdirs==2.0.2
prettytable==2.1.0
protobuf==3.15.8
psycopg2==2.9.5
py-gfm==0.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.20
pydicom==2.1.2
Pygments==2.1.3
PyNaCl==1.4.0
pyOpenSSL==22.0.0
PyPDF2==1.26.0
PyQt5==5.15.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.1
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.6.0
selenium==4.1.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
SpeechRecognition==3.10.0
SQLAlchemy==2.0.9
sqlparse==0.4.1
sudo==1.0.0
synapse-s3-storage-provider==1.1.2
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tqdm==4.64.1
traceback2==1.4.0
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.6.0
Twisted==22.10.0
typing_extensions==4.5.0
undetected-chromedriver==3.1.5.post4
unittest2==1.1.0
urllib3==1.26.4
virtualenv==20.6.0
watchdog==0.8.3
wcwidth==0.2.5
websocket-client==1.2.3
websocket-server==0.4
websockets==10.3
Werkzeug==2.2.3
wrapt==1.12.1
wsproto==1.1.0
XlsxWriter==1.3.7
yarl==1.9.1
zipp==3.15.0
zope.interface==5.5.2
